refactor(react-query): 폴더 구조 개선
