+++ apps/monitor-web/src/lib/index.ts
@@ -1,0 +1,2 @@
+export { supabase } from "./supabase";
+export { queryClient } from "./queryClient";
