+++ apps/monitor-web/src/lib/queryClient.ts
@@ -1,6 +1,6 @@
 import { QueryClient } from "@tanstack/react-query";
 
-const queryClient = new QueryClient({
+export const queryClient = new QueryClient({
   defaultOptions: {
     queries: {
       staleTime: 1000 * 60,
@@ -10,5 +10,3 @@ const queryClient = new QueryClient({
     },
   },
 });
-
-export default queryClient;
