+++ apps/monitor-web/src/main.tsx
@@ -2,8 +2,8 @@ import React from "react";
 import ReactDOM from "react-dom/client";
 import { BrowserRouter } from "react-router-dom";
 import { QueryClientProvider } from "@tanstack/react-query";
-import { queryClient } from "@/queries";
 import App from "./App";
+import { queryClient } from "./lib";
 import "./index.css";
 
 ReactDOM.createRoot(document.getElementById("root")!).render(
