+++ apps/monitor-web/src/queries/base/useAppQuery.ts
