+++ apps/monitor-web/src/queries/index.ts
@@ -1,5 +1,1 @@
-export { default as queryClient } from "./queryClient";
-export { default as useAppQuery } from "./useAppQuery";
-export { default as useAppMutation } from "./useAppMutation";
-export * from "./queryKeys";
-export * from "./apis.queries";
+export * from "./mock";
