+++ apps/monitor-web/src/queries/mock/index.ts
@@ -1,0 +1,1 @@
+export { useMockListQuery } from "./mock.queries";
